- Move the folder/app configuration logic into its own file
